backend/requirements.txt
# FastAPI Backend dependencies
fastapi
uvicorn
asyncpg
boto3
PyJWT[crypto]
upstash-redis
pydantic
python-multipart
httpx
